+++ HISTORY.md
@@ -17,6 +17,7 @@
 * objects: adjusted block-search-bar for variadic infix reporters
 * objects: adapted formula editor for variadic infix reporters
 * objects: removed scaffolding
+* store: apply block migration dictionary to hidden blocks in microworlds
 
 ### 2022-02-28
 * blocks, objects, threads, store: made addition reporter variadic